requirements.txt
aiosqlite==0.20.0
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
asgi-lifespan==2.1.0
asyncpg==0.30.0
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
iso8601==2.1.0
kombu==5.4.2
Mako==1.3.8
MarkupSafe==3.0.2
orjson==3.10.12
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
pypika-tortoise==0.3.2
pytest==8.3.4
pytest-asyncio==0.25.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tortoise-orm==0.22.2
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
